alphametics: Finish working draft.
Phew! At least the tests work after small modifications, it was cool
to get into println! debugging.  However, the last one especially is
too slow takes around 400s to run. I should try the following
optimisations:

- Use a column-major 1D array intead of Box<[Box<[Square]>]> (ugh,
never again!).

- Reduce allocations by making the possible moves into a groupoid, so
that we can mutate a global copy of the current Puzzle (probably
faster to undo a false move than to clone a Puzzle over methinks).

- Use data flow instead of control flow, think like an APLer!  Instead
of an array of Squares try an array of bytes with a separate array of
boolean flags answering the question "is it a digit or not?".