alphametics: AoS -> SoA for Puzzle data structure
Instead of a Puzzle being basically a Box<[Box<[Square]>]> we instead
represent it as a 2D column major array (represented using the Grid<T>
abstract data type).  I probably should format the carry row as well,
which isn't hard because the rest of the cells are one character long.