alphametics: Combine return value and arguments into one Grid.
The main reason I did this is because, I had contemplated a function
parse(words: Vec<Vec<u8>>) -> Result<Puzzle, E> but I might need to
pad both the arguments and the return value to have as many
cells (i.e. digits) as the return value, or even longer in the
degenerate case where a puzzle cannot be solved because one of the
arguments has more cells than the return value.  I could have solved
the problem by making the degenerate case into a parse error, so that
I can then pad every argument to match the return value but I think
the cache-locality benefit of not having a separate pointer into the
return value is useful as well.  Also, I don't think its quite _right_
to say that such a degenerate puzzle should give a _parse_ error, the
puzzle might be _syntactically_ valid but without solutions.