alphametics: Broken partial solution.
There's a problem with padding, essentially I've been assuming that
the beginning of each column is a variable but this isn't correct
because it could be a padding zero.  Now I've changed the code to
assume that you can skip padding at the _beginning_ of a column but
this isn't correct because the rows aren't sorted by length before
padding, so you can have chunks of padding anywhere in a column.  So,
I have to have three different types of Moves:

- Move::BindFresh which assumes that the cursor is on a variable and
then changes the Environment to include it.  Unlike the current
Move::Bind, it doesn't change the cursor.

- Move::Advance { row: usize } which moves the cursor down the column
skipping digits and substituting bound variables.  After the move is
done the cursor is on an unbound variable if one exists in the same
column below the row, or at the end of the column otherwise.

- Move::Carry { digit: u8, carry: u16 }