alphametics: Easy solution
Right so this "polynomial" idea works much better, 2.41 sec vs ~400 sec for the in-place backtracking!  I kinda have an idea why the backtracking is so slow, esp. since we don't have to do all 10! permutations as we would have to in the worst case.  I'd want to compare them though in that worst case to see, it might be that computing a bunch of dot products very fast is faster than even that sort of backtracking search though.